Make project IDE agnostic and remove Ant file
Previous state of project was linked to Eclipse by .classpath and
.project files and .settings directory and sub files.

There is no reason to keep it that way, since project was already
defined as IDE agnostic by the Ant configuration file.

In addition, the project has been migrated to Maven, so the Ant
configuration file can be removed too.

Eclipse specific files are removed from versioning and added to ignore
patterns (to prevent re-adding them by mistake) as the pom.xml and src
directory are enough to build the project from scratch.

This allows further contributors to use the IDE of their choosing, be it
Eclipse, Netbeans, Intellij IDEA or anything else. Please remember to
update the ignore patterns if needed.