Make OpenERPXmlRpcProxy use Proxy if defined
Proxy for HTTP and/or HTTPS is defined through system properties
http.proxyHost, http.proxyPort, https.proxyHost and https.proxyPort.

When available, the Proxy is used by setting its parameters in the
XmlRpcTransportFactory (available since Java 1.5 and later).

Also fix a bug in Session relative to synchrnoization of sessions
starting.