Remove getCommand introducing cyclic dependency
This method was just a convenience one, but created an uncalled for
cyclic dependency between Session and OpenERPCommand.

Besides, calling
OpenERPCommand command = session.getCommand();
instead of
OpenERPCommand command = new OpenERPCommand(session);
is not really that much simpler...

Also renamed attribute ObjectAdapter#commands to ObjectAdapter#command,
as it is a single instance, not a collection.