OdooCommand#callObjectFunction returns new class Response
Response wraps either a successful or failed call, indicated by
 #isSuccessful().
In case of success, actual response can be obtained "as is" with
 #getResponseObject() or as an array with #getResponseObjectAsArray().
The latter will either cast the response object as an array of objects if
it is an instance of one, and wrap the response object in a new array
otherwise.
In case of failure, the error cause can be accessed with
 #getErrorCause().

Also some code cleaning and refactoring.

Finally got to rename OpeneERPApiException to OdooApiException and
OpenERPCommand to OdooCommand.

This commit should provide the enhancement/fix requested in #7.